Refactored : Packages. Added:Shared Preference Class, Create note to Firebase
