+++ app/src/main/AndroidManifest.xml
@@ -9,15 +9,15 @@
         android:roundIcon="@mipmap/ic_launcher_round"
         android:supportsRtl="true"
         android:theme="@style/AppTheme.NoActionBar">
-        <activity android:name=".SplashActivity">
+        <activity android:name=".UI.Activity.SplashActivity">
             <intent-filter>
                 <action android:name="android.intent.action.MAIN" />
 
                 <category android:name="android.intent.category.LAUNCHER" />
             </intent-filter>
         </activity>
-        <activity android:name=".NotesActivity" />
-        <activity android:name=".MainActivity"/>
+        <activity android:name=".DashBoard.Activity.HomeActivity" />
+        <activity android:name=".UI.Activity.LoginRegisterActivity"/>
     </application>
 
 </manifest>
