+++ app/src/main/java/com/example/fundoonotes/DashBoard/Fragments/ArchiveFragment.java
@@ -1,4 +1,4 @@
-package com.example.fundoonotes;
+package com.example.fundoonotes.DashBoard.Fragments;
 
 import android.os.Bundle;
 
@@ -7,6 +7,8 @@ import androidx.fragment.app.Fragment;
 import android.view.LayoutInflater;
 import android.view.View;
 import android.view.ViewGroup;
+
+import com.example.fundoonotes.R;
 
 public class ArchiveFragment extends Fragment {
