+++ app/src/main/java/com/example/fundoonotes/DashBoard/Fragments/ReminderFragment.java
@@ -1,4 +1,4 @@
-package com.example.fundoonotes;
+package com.example.fundoonotes.DashBoard.Fragments;
 
 import android.os.Bundle;
 
@@ -8,22 +8,14 @@ import android.view.LayoutInflater;
 import android.view.View;
 import android.view.ViewGroup;
 
+import com.example.fundoonotes.R;
+
 public class ReminderFragment extends Fragment {
 
-    public ReminderFragment() {
-        // Required empty public constructor
-    }
-
-    @Override
-    public void onCreate(Bundle savedInstanceState) {
-        super.onCreate(savedInstanceState);
-
-    }
 
     @Override
     public View onCreateView(LayoutInflater inflater, ViewGroup container,
                              Bundle savedInstanceState) {
-        // Inflate the layout for this fragment
         return inflater.inflate(R.layout.fragment_reminder, container, false);
     }
 }
