+++ app/src/main/java/com/example/fundoonotes/DashBoard/Fragments/TrashFragment.java
@@ -1,4 +1,4 @@
-package com.example.fundoonotes;
+package com.example.fundoonotes.DashBoard.Fragments;
 
 import android.os.Bundle;
 
@@ -8,9 +8,11 @@ import android.view.LayoutInflater;
 import android.view.View;
 import android.view.ViewGroup;
 
-public class NotesFragment extends Fragment {
+import com.example.fundoonotes.R;
 
-    public NotesFragment() {
+public class TrashFragment extends Fragment {
+
+    public TrashFragment() {
         // Required empty public constructor
     }
 
@@ -24,6 +26,6 @@ public class NotesFragment extends Fragment {
     public View onCreateView(LayoutInflater inflater, ViewGroup container,
                              Bundle savedInstanceState) {
         // Inflate the layout for this fragment
-        return inflater.inflate(R.layout.fragment_notes, container, false);
+        return inflater.inflate(R.layout.fragment_trash, container, false);
     }
 }
