--- app/src/main/java/com/example/fundoonotes/MainViewModel.java
@@ -1,8 +1,0 @@
-package com.example.fundoonotes;
-
-import androidx.lifecycle.ViewModel;
-
-public class MainViewModel extends ViewModel {
-
-
-}
