--- app/src/main/java/com/example/fundoonotes/RegisterViewModel.java
@@ -1,7 +1,0 @@
-package com.example.fundoonotes;
-
-import androidx.lifecycle.ViewModel;
-
-public class RegisterViewModel extends ViewModel {
-    // TODO: Implement the ViewModel
-}
