+++ app/src/main/java/com/example/fundoonotes/UI/Activity/LoginRegisterActivity.java
@@ -1,31 +1,14 @@
-package com.example.fundoonotes;
+package com.example.fundoonotes.UI.Activity;
 
-import androidx.annotation.NonNull;
 import androidx.appcompat.app.AppCompatActivity;
-import androidx.fragment.app.Fragment;
-import androidx.fragment.app.FragmentManager;
-import androidx.fragment.app.FragmentTransaction;
 
-import android.content.Intent;
 import android.os.Bundle;
-import android.text.Editable;
-import android.text.TextUtils;
-import android.text.method.HideReturnsTransformationMethod;
-import android.text.method.PasswordTransformationMethod;
-import android.widget.Button;
-import android.widget.CheckBox;
-import android.widget.CompoundButton;
-import android.widget.EditText;
-import android.widget.FrameLayout;
-import android.widget.TextView;
-import android.widget.Toast;
 
-import com.google.android.gms.tasks.OnCompleteListener;
-import com.google.android.gms.tasks.Task;
-import com.google.firebase.auth.AuthResult;
-import com.google.firebase.auth.FirebaseAuth;
+import com.example.fundoonotes.R;
+import com.example.fundoonotes.UI.Fragments.RegisterFragment;
+import com.example.fundoonotes.UI.Fragments.LoginFragment;
 
-public class MainActivity extends AppCompatActivity {
+public class LoginRegisterActivity extends AppCompatActivity {
 
     private static final String LOGIN_FRAGMENT_TAG = "LoginFragment";
