+++ app/src/main/res/layout/activity_main.xml
@@ -5,7 +5,7 @@
     android:layout_width="match_parent"
     android:layout_height="match_parent"
     android:background="#F9FAF9"
-    tools:context=".MainActivity">
+    tools:context=".UI.Activity.LoginRegisterActivity">
 
     <FrameLayout
         android:id="@+id/fragment_container"
