+++ app/src/main/res/layout/navigation_header.xml
@@ -1,8 +1,8 @@
 <?xml version="1.0" encoding="utf-8"?>
 <LinearLayout xmlns:android="http://schemas.android.com/apk/res/android"
     android:layout_width="match_parent"
-    android:layout_height="176dp"
-    android:background="#D3FFE57F"
+    android:layout_height="180dp"
+    android:background="#F00C0C0C"
     android:gravity="bottom"
     android:orientation="vertical"
     android:padding="16dp"
@@ -16,6 +16,12 @@
     <TextView
         android:layout_width="wrap_content"
         android:layout_height="wrap_content"
-        android:text="@string/fundoo_notes"
+        android:text="@string/user_name"
+        android:textAppearance="@style/TextAppearance.AppCompat.Large" />
+
+    <TextView
+        android:layout_width="wrap_content"
+        android:layout_height="wrap_content"
+        android:text="@string/user_email"
         android:textAppearance="@style/TextAppearance.AppCompat.Large" />
 </LinearLayout>
